requirement.txt
django
Pillow
numpy
djangorestframework
